+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - New parameter `faker.locale` added to the dump configuration file
 - The following keywords are now forbidden in the `variables` param: `revoke`, `rename`, `lock`, `unlock`, `optimize`, `repair`
 - Replace double quotes by single quotes in SQL queries
+- Refactor Mysqldump extensions
 
 ## [2.1.0] - 2020-11-10
 [2.1.0]: https://github.com/Smile-SA/gdpr-dump/compare/2.0.3...2.1.0